requirements.txt
aws-cdk-lib==2.146.0
constructs>=10.0.0,<11.0.0
aws_cdk.lambda_layer_kubectl_v30==2.0.0
boto3>=1.34.131
PyYAML>=6.0.1